Fix dataset bugs in sequencer
The dataset would not be reloaded when the sequencer dialog (original) was closed, causing an empty body list. This is fixed by removing the Close button. For the neuprint sequencer dialog, the dataset was not set properly after initial selection. It is fixed by coordinating signals emitted from dataset combo box creation